requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.7.2
selenium==4.15.2
webdriver-manager==4.0.1
